Make "next" in zbx step through the complete execution of ldir and the other repeating instructions (lddr, cpir, cpdr, inir, indr, otir, otdr).
In xtrs 4.9d and earlier, "step" and "next" would both step entirely
over ldir and lddr because of optimized (but inaccurate) emulation of
those instructions, but I recently merged a change that disables the
optimization, making it tedious to single-step through code that uses
ldir/lddr.  This change helps compensate.