Partial cleanup of ROM loading
compile_rom is generalized a bit to deal better with ROMs that don't
start at address 0.  This avoids wasting 12KB of space when compiling
in the ESF extension ROM, which starts at 0x3000.

main.c is changed to work with the improvement, but doesn't take full
advantage; it continues to assume the ROMs are located where they are
supposed to be and will break silently if they don't.  I ran out of
time to do a more thorough cleanup.