Update sound documentation
Recommend aoss first, not padsp, as aoss should always work on an ALSA
system, but padsp is specific to PulseAudio.  Also I've noticed that
at least on Ubuntu 24.04 with PipeWire, the ossdp daemon doesn't work
with xtrs (it hangs and causes xtrs to hang too), but aoss does work.

A better solution for the future would be to update xtrs to be able to
use ALSA directly.  I finally got around to reading some of the ALSA
API documention recently and am realizing this might not be so much
work as I had thought.

Bonus fix: I noticed that (again, at least on Ubuntu 24.04), "groff
-man -Tpdf" produces bad output; for example, URLs (.UR macro) are
simply dropped.  I changed the Makefile to use -Tdvi and dvipdf
instead, which produces much better output.