Workaround for using a MODELA/III image as the ROM
These images expect to be running on a Model 4P, whose native boot ROM
sets 405bh to a nonzero value if the Break key is pressed.  Zero this
location upon reset to prevent spurious Break key recognition.

Hopefully this won't break some app that expects 405bh to be preserved
across reset!

Thanks to Jens Günther for alerting me to this fix in SDLTRS.